Change BOM list, Ender 3 switch is not safe for Trident
You cannot know how a CEE 7 plug  is plugged in. You have a 50% chance that wires that are connected to N on the PSU are running L1, even when the switch got toggled. Your heater bed may have live voltage when switched off, as the SSR also only switches one pole.